Move all important files on upper level and simplify folders navigation
