+++ packages/talker/lib/src/observer.dart
