--- packages/talker/lib/src/observers/observers.dart
@@ -1,1 +1,0 @@
-export 'talker_observer.dart';
