+++ packages/talker/lib/src/settings.dart
