--- packages/talker/lib/src/settings/settings.dart
@@ -1,1 +1,0 @@
-export 'talker_settings.dart';
