+++ packages/talker/lib/src/talker.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
-import 'package:talker/src/observers/observers_manager.dart';
+import 'package:talker/src/utils/observers_manager.dart';
+import 'package:talker/src/utils/utils.dart';
 import 'package:talker/talker.dart';
 
 /// Talker - advanced exception handling and logging
