+++ packages/talker/lib/talker.dart
@@ -3,8 +3,9 @@ library talker;
 export 'package:talker_logger/talker_logger.dart';
 
 export 'src/extensions/extensions.dart';
+export 'src/filter.dart';
 export 'src/models/models.dart';
-export 'src/observers/observers.dart';
-export 'src/settings/settings.dart';
+export 'src/observer.dart';
+export 'src/settings.dart';
 export 'src/talker.dart';
-export 'src/utils/utils.dart';
+export 'src/well_known_titles.dart';
